Re #11478 Make sure options are followed

+++ Code/Mantid/Framework/PythonInterface/plugins/algorithms/LiquidsReflectometryReduction.py
@@ -48,7 +48,7 @@ class LiquidsReflectometryReduction(PythonAlgorithm):
         self.declareProperty(IntArrayProperty("LowResNormAxisPixelRange", [115, 210],
                                               IntArrayLengthValidator(2), direction=Direction.Input),
                                               "Pixel range to use in the low resolution direction of the normalizaion run")
-        self.declareProperty(FloatArrayProperty("TOFRange", [9000., 23600.],
+        self.declareProperty(FloatArrayProperty("TOFRange", [0., 340000.],
                                                 FloatArrayLengthValidator(2), direction=Direction.Input),
                                                 "TOF range to use")
         self.declareProperty("TofRangeFlag", True,
@@ -139,27 +139,31 @@ class LiquidsReflectometryReduction(PythonAlgorithm):
                                          dataPeakRange, bck_request, dataBackRange)
 
         # ----- Normalization -------------------------------------------------
-        # Load normalization
-        ws_event_norm = LoadEventNexus("REF_L_%s" % normalizationRunNumber,
-                                       OutputWorkspace="REF_L_%s" % normalizationRunNumber)
-        crop_request = self.getProperty("LowResNormAxisPixelRangeFlag")
-        low_res_range = self.getProperty("LowResNormAxisPixelRange").value
-        bck_request = self.getProperty("SubtractNormBackground").value
-        norm_cropped = self.process_data(ws_event_norm, TOFrange,
-                                         crop_request, low_res_range,
-                                         normPeakRange, bck_request, normBackRange)
-        # Avoid leaving trash behind
-        AnalysisDataService.remove(str(ws_event_norm))
+        perform_normalization = self.getProperty("NormFlag").value
+        if perform_normalization:
+            # Load normalization
+            ws_event_norm = LoadEventNexus("REF_L_%s" % normalizationRunNumber,
+                                           OutputWorkspace="REF_L_%s" % normalizationRunNumber)
+            crop_request = self.getProperty("LowResNormAxisPixelRangeFlag")
+            low_res_range = self.getProperty("LowResNormAxisPixelRange").value
+            bck_request = self.getProperty("SubtractNormBackground").value
+            norm_cropped = self.process_data(ws_event_norm, TOFrange,
+                                             crop_request, low_res_range,
+                                             normPeakRange, bck_request, normBackRange)
+            # Avoid leaving trash behind
+            AnalysisDataService.remove(str(ws_event_norm))
+    
+            # Sum up the normalization peak
+            norm_summed = SumSpectra(InputWorkspace = norm_cropped)
 
-        # Sum up the normalization peak
-        norm_summed = SumSpectra(InputWorkspace = norm_cropped)
-
-        # Normalize the data
-        normalized_data = data_cropped / norm_summed
-        # Avoid leaving trash behind
-        AnalysisDataService.remove(str(data_cropped))
-        AnalysisDataService.remove(str(norm_cropped))
-        AnalysisDataService.remove(str(norm_summed))
+            # Normalize the data
+            normalized_data = data_cropped / norm_summed
+            # Avoid leaving trash behind
+            AnalysisDataService.remove(str(data_cropped))
+            AnalysisDataService.remove(str(norm_cropped))
+            AnalysisDataService.remove(str(norm_summed))
+        else:
+            normalized_data = data_cropped
 
         normalized_data = ConvertToPointData(InputWorkspace=normalized_data,
                                              OutputWorkspace=str(normalized_data))
@@ -449,10 +453,10 @@ class LiquidsReflectometryReduction(PythonAlgorithm):
                 scaling_factor_file = scaling_factor_files[0]
                 if not os.path.isfile(scaling_factor_file):
                     logger.error("Could not find scaling factor file %s" % scaling_factor_file)
-                    return
+                    return workspace
             else:
                 logger.error("Could not find scaling factor file %s" % scaling_factor_file)
-                return
+                return workspace
 
         # Get the incident medium
         incident_medium = self.getProperty("IncidentMediumSelected").value
